requirements.txt
absl-py==0.12.0
astunparse==1.6.3
cachetools==4.2.2
certifi==2020.12.5
chardet==4.0.0
cycler==0.10.0
decorator==4.4.2
easyocr==1.3.1
flatbuffers==1.12
gast==0.3.3
google-auth==1.30.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
idna==2.10
imageio==2.9.0
imutils==0.5.4
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.4
matplotlib==3.4.1
networkx==2.5.1
numpy==1.20.2
oauthlib==3.1.0
opencv-python==4.5.1.48
opt-einsum==3.3.0
pandas==1.2.4
Pillow==8.2.0
protobuf==3.15.8
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
PyQt6==6.0.3
PyQt6-Qt6==6.0.2
PyQt6-sip==13.0.1
pytesseract==0.3.7
python-bidi==0.4.2
python-dateutil==2.8.1
pytz==2021.1
PyWavelets==1.1.1
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-image==0.18.1
scipy==1.6.3
six==1.15.0
tensorboard==2.5.0
tensorboard-data-server==0.6.0
tensorboard-plugin-wit==1.8.0
tensorflow==2.4.1
tensorflow-estimator==2.4.0
termcolor==1.1.0
tifffile==2021.4.8
torch==1.8.1+cpu
torchaudio==0.8.1
torchvision==0.9.1+cpu
typing-extensions==3.7.4.3
urllib3==1.26.4
Werkzeug==1.0.1
wrapt==1.12.1
